Work around bug in variable declaration inside library module
re-lib.xqm previously declared $re:genres and used it in two functions.
The variable errored as unavailable on first call to those functions and
succeeded on subsequent ones. I don’t understand this behavior.

This commit works around the failure to process the variable declaration
before the first time it is used by changing it to a let statement
inside each of the functions that uses it.

This is a work-around, rather than a fix, because the same variable is
used in two places, and therefore should be declared only once and reused.